server_service/requirements.txt
fastapi 
uvicorn 
supabase-py 
httpx 
python-multipart
passlib[bcrypt]
python-jose[cryptography]